Validate request body when creating upload
It will now return a 400 response status, instead of a 500. An error
is represented in the response as e.g.

{
    "statusCode": 400,
    "message": [
        "junit_report_xml must be a string",
        "github_run_number must be a number conforming to the specified constraints"
    ],
    "error": "Bad Request"
}

(In terms of implementation, I tried passing {transform: true} to the
ValidationPipe initializer and it seemed to do nothing. Hence doing the
transforms manually using @Type. I didn’t want to go deep into
investigating.)